tung

+++ server/api/cart/[id].delete.ts
@@ -1,20 +1,55 @@
-export default defineEventHandler((event) => {
+import { getServerSession } from "#auth"
+
+
+export default defineEventHandler(async (event) => {
     const prisma = event.context.prisma
-    const id = event.context.params.id
+  const {id} = getRouterParams(event)
+  const data =  await getServerSession(event)
+  const user = data?.user
 
 
 
 
+ let updatedCart = await  prisma.cart.update({
+  where:{
+    userId:user.id
+  },
+data:{
+  products:{
+    connect:{
+      id:parseInt(id)
 
-    
- prisma.cart.delete({
-    where: { id: id },
-  });
+    },
+    update:{
+      where:{
+          id:parseInt(id)
+      },
+      data:{
+        quantity:{
+          decrement:1
+        }
+      }
+    }
+  }
+},
+include:{
+  products:{}
+}
+})
+
+if (updatedCart.products.length <= 0) {
 
 
+await prisma.cart.delete({
+  where:{
+    userId: user.id
+  }
+})
 
+}
 
-  return{success:true}
+  return updatedCart
+
